Move class-like directives to separate file
